浮動小数点ではdecimalを使用するように変更 (#30)
* doubleからdecimalに変更

* ユニットテスト更新

* decimal.Parseが失敗していたので修正

+++ BitbankDotNet.Tests/PrivateApis/BitbankRestApiClientGetOrderAsyncTest.cs
@@ -38,15 +38,15 @@ namespace BitbankDotNet.Tests.PrivateApis
                 var result = bitbank.GetOrderAsync(default, default).GetAwaiter().GetResult();
 
                 Assert.NotNull(result);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.AveragePrice);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.ExecutedAmount);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.AveragePrice);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.ExecutedAmount);
                 Assert.Equal(EntityHelper.GetTestValue<DateTime>(), result.OrderedAt);
                 Assert.Equal(EntityHelper.GetTestValue<long>(), result.OrderId);
                 Assert.Equal(EntityHelper.GetTestValue<CurrencyPair>(), result.Pair);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Price);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.RemainingAmount);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Price);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.RemainingAmount);
                 Assert.Equal(EntityHelper.GetTestValue<OrderSide>(), result.Side);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.StartAmount);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.StartAmount);
                 Assert.Equal(EntityHelper.GetTestValue<OrderStatus>(), result.Status);
                 Assert.Equal(EntityHelper.GetTestValue<OrderType>(), result.Type);
             }
