+++ BitbankDotNet.Tests/PublicApis/BitbankRestApiClientGetTickerAsyncTest.cs
@@ -38,13 +38,13 @@ namespace BitbankDotNet.Tests.PublicApis
                 var result = bitbank.GetTickerAsync(default).GetAwaiter().GetResult();
 
                 Assert.NotNull(result);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Buy);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.High);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Last);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Low);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Sell);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Buy);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.High);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Last);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Low);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Sell);
                 Assert.Equal(EntityHelper.GetTestValue<DateTime>(), result.Timestamp);
-                Assert.Equal(EntityHelper.GetTestValue<double>(), result.Vol);
+                Assert.Equal(EntityHelper.GetTestValue<decimal>(), result.Vol);
             }
         }
