+++ BitbankDotNet/Entities/Transaction.cs
@@ -24,12 +24,12 @@ namespace BitbankDotNet.Entities
         /// <summary>
         /// 価格
         /// </summary>
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         /// <summary>
         /// 数量
         /// </summary>
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         /// <summary>
         /// 約定日時
